improves testing
