+++ test_scripts/data/ascii.garbage
@@ -1,0 +1,1 @@
+not emojisV2
