+++ test_scripts/data/concat_v1_1.enc
@@ -1,0 +1,1 @@
+👖📸🎈☕🎥🤠📠🏍🐲👡🕟☕
