+++ test_scripts/data/concat_v1_1.plaind
@@ -1,0 +1,1 @@
+abc6789XY
