+++ test_scripts/data/concat_v2_1.enc
@@ -1,0 +1,1 @@
+👖📸🧈🌭👩☕💲🥇🪚☕
