+++ test_scripts/data/concat_v2_1.plaind
@@ -1,0 +1,1 @@
+abcdefxyz
