+++ test_scripts/data/mixed_1.garbage
@@ -1,0 +1,1 @@
+🀄🅰🤿🤾
