+++ test_scripts/data/mixed_2.garbage
@@ -1,0 +1,1 @@
+🀄🤿🅰🤾
