+++ test_scripts/data/mixed_3.garbage
@@ -1,0 +1,1 @@
+🀄🅰🅰🤾🀄🅰🅰🤾🀄🅰🤿🤾
